Added \mathopen and \mathclose to \DeclareMathDelimiter
Reimplemented \DeclareMathDelimiterX to provide a better interface for
the user, plus rewrote some of the docs for it.

Note that the addition of \mathopen and \mathclose is NOT mentioned in
the manual (not officially).

Would it be an idea to provide a configuration option to disable the
addition of these \mathopen and \mathclose pairs?




git-svn-id: http://subversion.assembla.com/svn/mhctan@93 3f4560be-fcb6-4516-b323-ae570ede3bb7