provided fenced smallmatrix environments courtesy of Rasmus Villemoes
The environments all support the same syntax as matrix*, and thus have
all been named ending with a *. In addition the default alignment of
these small matrixes can be configured globally.

Rasmus have suggested a similar feature for the Xmatrix* environments,
but that need a little more work or it will collide with the array
package.

I also bumped the version number of mathtools

BTW: might be better if the lppl statement actually state who is
maintaining the packages, otherwise people will keep addressing the
mhctan list with Morten Hogholm.



git-svn-id: http://subversion.assembla.com/svn/mhctan@103 3f4560be-fcb6-4516-b323-ae570ede3bb7